+++ idb_shim/lib/idb.dart
@@ -6,43 +6,23 @@ import 'package:idb_shim/src/common/common_factory.dart';
 import 'package:idb_shim/src/common/common_key_range.dart';
 
 export 'src/client/error.dart';
-
-/// Read-write mode for transaction.
-const String idbModeReadWrite = 'readwrite';
-
-/// Read-only mode for transaction.
-const String idbModeReadOnly = 'readonly';
-
-/// Default forward mode for cursor.
-const String idbDirectionNext = 'next';
-
-/// Backward mode for cursor.
-const String idbDirectionPrev = 'prev';
-
-/// Factory name using native indexeddb implementation.
-const idbFactoryNameNative = 'native';
-
-/// Factory name using Sembast implementation
-const idbFactoryNameSembastIo = 'sembast_io';
-
-/// Factory name using Sembast io implementation.
-@Deprecated('Use idbFactoryNameSembastIo instead')
-const idbFactoryNameIo = 'io';
-
-/// Factory name using Sembast memory implementation
-const idbFactoryNameSembastMemory = 'sembast_memory';
-
-/// Factory name that could be used to use Sembast Memory implementation.
-const idbFactoryNameMemory = 'memory';
-
-/// Pseudo - best persistent shim (indexeddb).
-const idbFactoryNamePersistent = 'persistent';
-
-/// Pseudo - best browser shim (persistent of it not available memory).
-const idbFactoryNameBrowser = 'browser';
-
-/// Shim using WebSql implementation - no longer supported.
-const idbFactoryNameWebSql = 'websql';
+export 'src/constant.dart'
+    show
+        idbModeReadOnly,
+        idbModeReadWrite,
+        idbDirectionNext,
+        idbDirectionPrev,
+        idbFactoryNameBrowser,
+        idbFactoryNameLogger,
+        idbFactoryNameMemory,
+        idbFactoryNameNative,
+        idbFactoryNamePersistent,
+        idbFactoryNameSembastIo,
+        idbFactoryNameSembastMemory,
+        idbFactoryNameWebSql,
+        // ignore: deprecated_member_use_from_same_package
+        idbFactoryNameIo;
+export 'src/database_exception.dart' show DatabaseException;
 
 ///
 /// represents a cursor for traversing or iterating over multiple records in a
@@ -160,6 +140,10 @@ abstract class Transaction {
   /// complete when the transaction is done
   ///
   Future<Database> get completed;
+
+  /// Rolls back all the changes to objects in the database associated
+  /// with this transaction.
+  void abort();
 }
 
 ///
@@ -551,7 +535,8 @@ abstract class VersionChangeEvent {
 ///
 abstract class Event {}
 
-typedef OnUpgradeNeededFunction = void Function(VersionChangeEvent event);
+typedef OnUpgradeNeededFunction = FutureOr<void> Function(
+    VersionChangeEvent event);
 
 typedef OnBlockedFunction = void Function(Event event);
 
@@ -688,24 +673,4 @@ class DatabaseError extends Error {
 
   @override
   String toString() => message;
-}
-
-///
-/// Generic database exception.
-///
-class DatabaseException implements Exception {
-  /// Error message.
-  String get message => _message;
-  final String _message;
-
-  /// Create a database exception with a message.
-  DatabaseException(this._message);
-
-  @override
-  String toString() {
-    if (message == null) {
-      return 'DatabaseException';
-    }
-    return 'DatabaseException: $message';
-  }
 }