+++ idb_shim/lib/src/common/common_exception.dart
@@ -1,0 +1,9 @@
+import 'package:idb_shim/idb.dart';
+
+/// New abort exception to avoid creating them everywhere
+DatabaseException newAbortException([String message = 'Aborted']) =>
+    DatabaseException(message);
+
+/// New abort exception to avoid creating them everywhere
+DatabaseException wrapException(Exception e, [String message]) =>
+    DatabaseException('${e.runtimeType}: message');