+++ idb_shim/lib/src/sembast/sembast_database.dart
@@ -67,9 +67,11 @@ class DatabaseSembast extends IdbDatabaseBase with DatabaseWithMetaMixin {
       IdbFactory factory, sdb.Database db) async {
     final idbDb = DatabaseSembast._(factory);
     idbDb.db = db;
-    await idbDb._readMeta();
+    var version = await idbDb._readMeta();
     // Copy name from path
     idbDb.meta.name = db.path;
+    // devPrint('fromDatabase version $version meta ${idbDb.meta.version}');
+    idbDb.meta.version = version ?? 1;
     return idbDb;
   }
 
@@ -97,9 +99,11 @@ class DatabaseSembast extends IdbDatabaseBase with DatabaseWithMetaMixin {
   // return the previous version
   Future<int> _readMeta() async {
     return db.transaction((txn) async {
+      // devPrint(await mainStore.find(txn));
+
       // read version
       meta.version = await mainStore.record('version').get(txn) as int;
-      //devPrint('meta version :${meta.version})
+      // devPrint('read meta version ${meta.version} ${db.path}');
       // read store meta
       var storeList = await mainStore.record('stores').get(txn);
       if (storeList != null) {
@@ -126,52 +130,73 @@ class DatabaseSembast extends IdbDatabaseBase with DatabaseWithMetaMixin {
           'open2 $onUpgradeNeeded ${onUpgradeNeeded != null ? 'NOT NULL' : 'NULL'}');
     }
     // Open the sembast database
-    db = await sdbFactory.openDatabase(factory.getDbPath(name), version: 1);
-    previousVersion = await _readMeta();
-    if (newVersion != previousVersion) {
-      Set<IdbObjectStoreMeta> changedStores;
-      Set<IdbObjectStoreMeta> deletedStores;
+    db = await sdbFactory.openDatabase(factory.getDbPath(name), version: 1,
+        onVersionChanged: (db, oldVersion, newVersion) {
+      if (sembastDebug) {
+        print('changing ${db.path} $oldVersion -> $newVersion');
+      }
+    });
+    try {
+      previousVersion = await _readMeta();
+      // devPrint('Opening $name previous $previousVersion new $newVersion version $version');
+      // If not specify and previous version is null, force new version to 1
+      if (previousVersion == null) {
+        newVersion ??= 1;
+      }
+      if (newVersion != null && newVersion != previousVersion) {
+        Set<IdbObjectStoreMeta> changedStores;
+        Set<IdbObjectStoreMeta> deletedStores;
 
-      await meta.onUpgradeNeeded(() async {
-        versionChangeTransaction =
-            TransactionSembast(this, meta.versionChangeTransaction);
-        // could be null when opening an empty database
-        if (onUpgradeNeeded != null) {
-          onUpgradeNeeded(
-              _SdbVersionChangeEvent(this, previousVersion, newVersion));
-        }
-        changedStores = Set.from(meta.versionChangeTransaction.createdStores);
-        changedStores.addAll(meta.versionChangeTransaction.updatedStores);
-        deletedStores = meta.versionChangeTransaction.deletedStores;
-      });
+        await meta.onUpgradeNeeded(() async {
+          versionChangeTransaction =
+              TransactionSembast(this, meta.versionChangeTransaction);
+          // could be null when opening an empty database
+          if (onUpgradeNeeded != null) {
+            await onUpgradeNeeded(
+                _SdbVersionChangeEvent(this, previousVersion, newVersion));
+          }
 
-      await db.transaction((txn) async {
-        await mainStore.record('version').put(txn, newVersion);
+          await versionChangeTransaction.completed;
+          changedStores = Set.from(meta.versionChangeTransaction.createdStores);
+          changedStores.addAll(meta.versionChangeTransaction.updatedStores);
+          deletedStores = meta.versionChangeTransaction.deletedStores;
+        });
 
-        // First delete everything from deleted stores
-        for (final storeMeta in deletedStores) {
-          await sdb.StoreRef(storeMeta.name).drop(txn);
-        }
+        await db.transaction((txn) async {
+          await mainStore.record('version').put(txn, newVersion);
 
-        // Handle deleted object store
-        if (changedStores.isNotEmpty || deletedStores.isNotEmpty) {
-          await mainStore.record('stores').put(
-              txn,
-              List.from(objectStoreNames)
-                ..sort()); // Sort the names to always have the same export
-        }
+          // First delete everything from deleted stores
+          for (final storeMeta in deletedStores) {
+            await sdb.StoreRef(storeMeta.name).drop(txn);
+          }
 
-        for (final storeMeta in changedStores) {
-          await mainStore
-              .record('store_${storeMeta.name}')
-              .put(txn, storeMeta.toMap());
-        }
-      }).then((_) {
+          // Handle deleted object store
+          if (changedStores.isNotEmpty || deletedStores.isNotEmpty) {
+            await mainStore.record('stores').put(
+                txn,
+                List.from(objectStoreNames)
+                  ..sort()); // Sort the names to always have the same export
+          }
+
+          for (final storeMeta in changedStores) {
+            await mainStore
+                .record('store_${storeMeta.name}')
+                .put(txn, storeMeta.toMap());
+          }
+        });
         // considered as opened
         meta.version = newVersion;
-      });
+      } else {
+        // Keep existing meta
+        meta.version = previousVersion ?? 1;
+      }
+      return db;
+    } catch (e) {
+      try {
+        await db?.close();
+      } catch (_) {}
+      rethrow;
     }
-    return db;
   }
 
   @override