+++ idb_shim/CHANGELOG.md
@@ -1,6 +1,12 @@
-## 1.12.0
+## 1.12.2
+
+* Add support for `Transaction.abort`
+* Allow read/write during open transaction.
+
+## 1.12.1+1
 
 * Add `ObjectStore.getAll/getAllKeys` and `Index.getAll/getAllKeys`
+* Fix import/export through sembast
 
 ## 1.11.1+1